requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.8.3
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2